Create button for node reallocation
Adds a button to the Lease Overview tab which allows the user to reallocate any of the hosts allocated to their lease. Also shows the "status" of each node: healthy/unhealthy, which maps to reservable/unreservable, or "unknown" if "reservable" is not set.

In order to embed the buttons in the Lease Overview tab, a patch to Horizon had to be made: https://github.com/ChameleonCloud/horizon/pull/18

The front-end changes were a little complicated. Some extra tag parameters needed to be added in order to get all of the dialogue functioning properly. Horizon would normally inject this information for you when buttons are defined as python classes, but in this case, that was not possible.