Fix datepicker check in lease creation workflow
The callback for the "number of days" selector checked the `startDate`
datepicker for "Invalid Date," which is only returned for an input which
does not resolve to a valid date. If the datepicker has not been set at
all, it returns `null`. The callback function now also checks for
`null`, to indicate that a start date has not been specified. This way,
it will use today's date instead of the beginning of the Unix epoch.