Update tox.ini for tox4
OpenStack CI now uses tox4 by default. This patch
drop "skipsdist = True". When skipsdist=True is specified,
a target application is not installed into a tox env, but
there seems no reason to do so. In various other OpenStack
projects, a target applicatin needs to be installed into a
tox venv, for example, in docs and/or linter jobs and it
causes failures with tox4. So dropping it would avoid potential
future failures.

Change-Id: I18d5ee7f96e77e345f304536bbc385d87bb859f2