Fix network reservation calendar
Both the resource and reservation treated segment_id as an integer,
which was only a problem because we were converting the
resource.segment_id to a string. This meant they weren't being matched
when generating the calendar rows, and so any resource with reservations
showed up twice, one for the string segment_id, one for the int
segment_id.